+++ src/components/GNotification/index.ts
@@ -1,0 +1,6 @@
+import GNotificationContainer from './GNotificationContainer.tsx';
+import useNotification from './useNotification.ts';
+
+export { useNotification };
+
+export default GNotificationContainer;